// src/screens/trades/types.ts
export type TTradesCreate = {
  trade_date: any;
  product: string;
  customer_ref: string;
  acc_name: string;
  acc_no: string;
  branch: string;
  acc_balance: number;
  acc_currency: string;
  comment: string;
  authorizer_comment: string;
  direction: string;
  direction_ccy: string;
  we_sell: number;
  we_sell_ccy: string;
  exchange_rate1: number;
  sell_settle_date: any;
  debit_amount: number;
  debit_ccy: string;
  exchange_rate2: number;
  customer_debit_settle_date: any;
  reference_ccy_settle_date: any;
  ref_amount: number;
  ref_ccy: string;
  transfer_purpose: string;
  form_a_no: string;
  passport_no: string;
  travel_country: string;
  e_ticket_no: string;
  application_no: string;
  travel_date: any;
  allocation_date: any;
  status: any;
  dorm_accno: string;
  nostro_acct: number;
  dorm_accname: string;
  dormacc_currency: string;
};
export type TTradesEdit = {};
export type TTradesView = {};
export type TTradesList = {
  trade_date: any;
  product: string;
  customer_ref: string;
  acc_name: string;
  acc_no: string;
  branch: string;
  acc_balance: number;
  acc_currency: string;
  comment: string;
  authorizer_comment: string;
  direction: string;
  direction_ccy: string;
  we_sell: number;
  we_sell_ccy: string;
  exchange_rate1: number;
  sell_settle_date: any;
  debit_amount: number;
  debit_ccy: string;
  exchange_rate2: number;
  customer_debit_settle_date: any;
  reference_ccy_settle_date: any;
  ref_amount: number;
  ref_ccy: string;
  transfer_purpose: string;
  form_a_no: string;
  passport_no: string;
  travel_country: string;
  e_ticket_no: string;
  application_no: string;
  travel_date: any;
  allocation_date: any;
  status: any;
  dorm_accno: string;
  dorm_accname: string;
  dormacc_currency: string;
};